Load Bangalore Map, drew a polygon for Bel campus
Loading Bengaluru map: Loaded a detailed map of BENGALURU, focusing on the region around Bharat Electronics Limited.

Drawing Polygon: Created a polygon to highlight the BHEL area on the map. It includes loading tile imagery from a server, defining the coordinates of the BHEL campus, and controlling the camera view.